Fix the wrong mask for bit write
When writing bits to the destination without crossing mulitple
bytes, we need to clear the desired bits that we want to write and
reserve other bits by executing the bitwise AND with 1'.

However, the original logic does not adere to the above-mentioned
statement. This leads to the unexpected result when the bits of the
variables 'input' and 'output' are set to 0/1.